clippy

+++ framework/scenario/src/debug_executor/tx_static_vars.rs
@@ -8,7 +8,7 @@ pub struct TxStaticVars {
     pub call_value_egld_handle: RawHandle,
     pub call_value_multi_esdt_handle: RawHandle,
     //vec of true/false, true if bit from handle = scaling_start + index is not empty
-    pub scaling_factor_init: [bool; const_handles::SCALING_FACTOR_LENGTH as usize],
+    pub scaling_factor_init: [bool; const_handles::SCALING_FACTOR_LENGTH],
 }
 
 impl Default for TxStaticVars {
@@ -19,7 +19,7 @@ impl Default for TxStaticVars {
             num_arguments: -1,
             call_value_egld_handle: const_handles::UNINITIALIZED_HANDLE,
             call_value_multi_esdt_handle: const_handles::UNINITIALIZED_HANDLE,
-            scaling_factor_init: [false; const_handles::SCALING_FACTOR_LENGTH as usize],
+            scaling_factor_init: [false; const_handles::SCALING_FACTOR_LENGTH],
         }
     }
 }
